provider/common: ensure size_t-correctness for *Size functions

+++ provider/common/SOAPUtils.h
@@ -58,7 +58,7 @@ extern ECRESULT MergePropValArray(struct soap *soap, const struct propValArray *
 extern ECRESULT CompareProp(const struct propVal *lpProp1, const struct propVal *lpProp2, const ECLocale &locale, int *lpCompareResult);
 extern ECRESULT CompareMVPropWithProp(struct propVal *lpMVProp1, const struct propVal *lpProp2, unsigned int ulType, const ECLocale &locale, bool* lpfMatch);
 
-unsigned int		PropSize(const struct propVal *);
+size_t PropSize(const struct propVal *);
 ECRESULT			FreePropVal(struct propVal *lpProp, bool bBasePointerDel);
 ECRESULT CopyPropVal(const struct propVal *lpSrc, struct propVal *lpDst, struct soap *soap = NULL, bool bTruncate = false);
 ECRESULT CopyPropVal(const struct propVal *lpSrc, struct propVal **lppDst, struct soap *soap = NULL, bool bTruncate = false); /* allocates new lpDst and calls other version */
@@ -104,14 +104,14 @@ ULONG 				NormalizePropTag(ULONG ulPropTag);
 
 const char *GetSourceAddr(struct soap *soap);
 
-unsigned int SearchCriteriaSize(const struct searchCriteria *);
-unsigned int RestrictTableSize(const struct restrictTable *);
-unsigned int PropValArraySize(const struct propValArray *);
-unsigned int EntryListSize(const struct entryList *);
-unsigned int EntryIdSize(const entryId *);
-unsigned int NotificationStructSize(const notification *);
-unsigned int PropTagArraySize(const struct propTagArray *);
-unsigned int SortOrderArraySize(const struct sortOrderArray *);
+size_t SearchCriteriaSize(const struct searchCriteria *);
+size_t RestrictTableSize(const struct restrictTable *);
+size_t PropValArraySize(const struct propValArray *);
+size_t EntryListSize(const struct entryList *);
+size_t EntryIdSize(const entryId *);
+size_t NotificationStructSize(const notification *);
+size_t PropTagArraySize(const struct propTagArray *);
+size_t SortOrderArraySize(const struct sortOrderArray *);
 
 class DynamicPropValArray {
 public:
